script/test-requirements.txt
nose2==0.6.2
mock==2.0.0
PyHamcrest==1.9.0
pytest
